+++ src/2-two/postTagger/model/nouns/nouns.js
@@ -138,7 +138,7 @@ export default [
   // a representative to 
   { match: '#Determiner [%Adj|Noun%] #Conjunction', group: 0, tag: 'Noun', reason: 'a-rep-to' },
   // near death experiences, ambitious sales targets
-  { match: '#Adjective #Noun [%Plural|Verb%]$', group: 0, tag: 'Plural', reason: 'near-death-experiences' },
+  { match: '#Adjective #Noun [%Plural|Verb%]$', group: 0, tag: 'Plural', notIf: '#Pronoun', reason: 'near-death-experiences' },
   // your guild colors
   { match: '#Possessive #Noun [%Plural|Verb%]$', group: 0, tag: 'Plural', reason: 'your-guild-colors' },